requirements.txt
accelerate==0.27.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anthropic==0.15.1
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.4
argcomplete @ file:///private/tmp/python-argcomplete-20240123-4975-yeh326/argcomplete-3.2.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==1.4.8
asgiref==3.7.2
asttokens==2.4.1
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Authlib==1.3.0
Automat==22.10.0
Babel==2.14.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.22.1
bitsandbytes==0.42.0
black==23.3.0
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
bpyutils==0.5.8
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
clickhouse-connect==0.7.0
cloudpathlib==0.16.0
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.4
constantly==15.1.0
contourpy==1.2.0
cryptography==40.0.2
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
datasets==2.17.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
dnspython==2.5.0
docopt==0.6.2
duckdb==0.9.2
elastic-transport==8.12.0
elasticsearch==8.12.0
entrypoints==0.4
eventlet==0.35.1
exceptiongroup==1.2.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.109.2
fastjsonschema==2.19.1
feedparser==6.0.11
ffmpy==0.3.2
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.48.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.6.0
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-search-results==2.4.2
googleapis-common-protos==1.62.0
gotrue==1.3.0
gradio==3.28.3
gradio_client==0.10.0
greenlet==2.0.2
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.60.1
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hnswlib==0.8.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.20.3
humanfriendly==10.0
Hypercorn==0.16.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.6.0
importlib-resources==6.1.1
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.21.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.5
langchainplus-sdk==0.0.20
langcodes==3.3.0
langsmith==0.0.90
linkify-it-py==2.0.3
loguru==0.7.2
loralib==0.1.2
lxml==4.9.3
lz4==4.3.3
manifest-ml==0.1.9
markdown-it-py==2.2.0
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-enum==1.5.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==2.0.5
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook_shim==0.2.3
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.0
openai==1.12.0
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.13
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.3
pathlib_abc==0.1.1
pathspec==0.12.1
pathy==0.11.0
peft==0.8.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
pipreqs==0.4.13
pipupgrade==1.12.0
platformdirs==3.5.1
pluggy==1.4.0
postgrest==0.15.0
posthog==3.4.1
preshed==3.0.9
priority==2.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==12.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.2
pydeck==0.8.1b0
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.17.2
pyOpenSSL==23.1.1
pyparsing==3.1.1
pypdf==3.9.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyrsistent==0.20.0
PySocks==1.7.1
pytest==7.3.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-json-logger==2.0.7
python-multipart==0.0.9
python-socketio==5.11.1
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
Quart==0.19.4
quart-cors==0.7.0
queuelib==1.6.2
realtime==1.0.2
redis==5.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
responses==0.24.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scipy==1.12.0
Scrapy==2.11.0
selenium==4.17.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
service-identity==21.1.0
sgmllib3k==1.0.0
simple-websocket==0.10.0
simsimd==3.7.6
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.26
sqlitedict==2.1.0
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.1
storage3==0.7.0
streamlit==1.31.1
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.0
thinc==8.2.3
tiktoken==0.6.0
tinycss2==1.2.1
tldextract==3.4.1
tokenize-rt==5.2.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
trio==0.24.0
trio-websocket==0.11.1
Twisted==22.10.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.0
uvicorn==0.27.1
uvloop==0.19.0
validators==0.22.0
w3lib==2.1.2
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
weaviate-client==2.5.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
zstandard==0.22.0
